Moved WorkWithUs to the root Compositions folder and added it to the services page as well
