+++ src/compositions/WorkWithUs/spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import WorkWithUs from './index';
 
-test('compositions/homepage/WorkWithUs snapshot', () => {
+test('compositions/WorkWithUs snapshot', () => {
   const wrapper = shallow(<WorkWithUs />);
 
   expect(wrapper).toMatchSnapshot();
