+++ src/compositions/pages/Home/index.jsx
@@ -1,5 +1,7 @@
 import React, { Fragment } from 'react';
 
+import WorkWithUs from 'compositions/WorkWithUs';
+
 import BuildBetter from './BuildBetter';
 import CaseStudies from './CaseStudies';
 import Industries from './Industries';
@@ -7,7 +9,6 @@ import WeEnable from './WeEnable';
 import HowWeDoIt from './HowWeDoIt';
 import Community from './Community';
 import ServicesAboutUs from './ServicesAboutUs';
-import WorkWithUs from './WorkWithUs';
 
 const HomePage = () => (
   <Fragment>
